+++ client/package.json
@@ -10,6 +10,7 @@
     "bootstrap": "^4.5.3",
     "lorem-ipsum": "^2.0.3",
     "mdbreact": "^4.27.0",
+    "pg": "^8.5.1",
     "react": "^17.0.1",
     "react-bootstrap": "^1.4.0",
     "react-dom": "^17.0.1",
@@ -17,6 +18,7 @@
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.0",
     "redux": "^4.0.5",
+    "sequelize-cli": "^6.2.0",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
