+++ client/src/styles/projects.css
@@ -1,0 +1,30 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+#top {
+    width: 100vw;
+    min-height: 100vh;
+    background-image: linear-gradient(-225deg, #2CD8D5 0%, #C5C1FF 56%, #FFBAC3 100%);
+    position: relative;
+    display: flex;
+}
+
+button.participate-button {
+    margin-top: 15px;
+    width: 100%;
+    height: 50px;
+    border: 0;
+    cursor: pointer;
+    background: #3cdc8c;
+    /* border-radius: 20px; */
+    color: #ffffff;
+    font-weight: 800;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    transition: background-color 0.2s;
+    font-size: 1.2em;
+}
